Convert MultiPolygon to Polygon for ST connector
This update was made because the FROST server no longer accepts
MultiPolygons as valid WKT inputs. This change will allow the
ST connector to convert MultiPolygons to Polygons before sending
them to the FROST server. However, if the MultiPolygon is not
continuous, then the MultiPolygon won't be converted to a Polygon