Separate analyte and water level sources to differentiate where data comes from in config report
There are some sources that only report water levels or analyte
measurements. Prior to this update, for example, if the user requested
water level data, the sources that only reported analyte measurements
would say that data was being pulled from them. Now, the sources that
only report water level or analyte measurements are separated so that
the user can see where the data is coming from.