Updated source names | Set source to False if not in use
Source names have been updated to indicate the organization and not
where the data is hosted (such as CKAN or ST2). State agencies have
also been prepended to OSE Roswell, ISC Seven Rivers, AMP, and DWB for
clarity.

Set source to False if the source is not in use. With this change, however,
the user can still set `--no-{source}` even if it doesn't report that
parameter. This will prevent unecessary crashes