Put each output for each call into unique output directory
This is done so that outputs don't override each other. All outputs are
now found in directories titled `output`. If the output directory already
exists, the output will be placed in a directory with the same name but
with a number appended to the end. For example, if the output directory
`output` already exists, the output will be placed in `output_1`. If
`output_1` already exists, the output will be placed in `output_2`, and so
on.

die.log is also placed in the output directory so that each tool call
is recorded in a single log file.