Address pr review comments about blocking unit conversions
Unit conversions are now blocked by output_units where possible (unless
the analyte is pH, in which case the die_parameter_name is used). All
conversions for output units are now assessed in if/elif statements for
clarity and ease. Furthermore, no else statements are used so that if
the conversion is not accounted for it will be reported to the user